size_t -> unsigned for best_hyps

+++ src/amun/gpu/decoder/best_hyps.h
@@ -32,7 +32,7 @@ class BestHyps : public BestHypsBase
                    const bool isFirst);
 
     std::vector<SoftAlignmentPtr> GetAlignments(const std::vector<ScorerPtr>& scorers,
-                                                size_t hypIndex);
+                                                unsigned hypIndex);
 
     void CalcBeam(
         const Beam& prevHyps,
